+++ R/beast.R
@@ -68,14 +68,11 @@ read.treetext_beast <- function(file) {
     jj <- jj[jj > max(ii)][1]
     jj <- c(ii[-1], jj)
 
-    trees <- sapply(seq_along(ii), function(i) {
+    trees <- lapply(seq_along(ii), function(i) {
         tree <- beast[(ii[i]+1):(jj[i]-1)]
         tree <- tree[grep("^\\s*tree", tree, ignore.case = TRUE)]
-        ## if (length(tree) > 1) {
-        ##     tree <- paste0(tree, collapse='')
-        ## }
         sub("[^(]*", "", tree)
-    })
+    }) %>% unlist
 
     return(trees)
 }
@@ -94,8 +91,8 @@ read.trans_beast <- function(file) {
     trans %<>% `[`(nzchar(trans))
     ## remove quote if strings were quoted
     trans %<>% gsub("'|\"", "",.)
-    trans %<>% sapply(., strsplit, split="\\s+")
-    trans %<>% do.call(rbind, .)
+    trans <- strsplit(trans, split="\\s+") %>%
+        do.call(rbind, .)
     ## trans is a matrix
     return(trans)
 }
@@ -129,48 +126,6 @@ read.stats_beast_internal <- function(beast, tree) {
     root <- rootnode(phylo)
     nnode <- phylo$Nnode
 
-    ## phylo2 <- read.nexus(file)
-    ## treeinfo <- fortify.phylo(phylo)
-    ## treeinfo2 <- fortify.phylo(phylo2)
-    ## treeinfo$label2 <- NA
-    ## treeinfo$label2[treeinfo$isTip] <- treeinfo2$node[as.numeric(treeinfo$label[treeinfo$isTip])]
-    ## treeinfo$visited <- FALSE
-    ## root <- getRoot(phylo2)
-    ## treeinfo[root, "visited"] <- TRUE
-    ## currentNode <- 1:Ntip(phylo2)
-    ## while(any(treeinfo$visited == FALSE)) {
-    ##     pNode <- c()
-    ##     for (kk in currentNode) {
-    ##         i <- which(treeinfo$label2 == kk)
-    ##         treeinfo[i, "visited"] <- TRUE
-    ##         j <- which(treeinfo2$node == kk)
-    ##         ip <- treeinfo$parent[i]
-    ##         if (ip != root) {
-    ##             ii <- which(treeinfo$node == ip)
-    ##             if (treeinfo$visited[ii] == FALSE) {
-    ##                 jp <- treeinfo2$parent[j]
-    ##                 jj <- which(treeinfo2$node == jp)
-    ##                 treeinfo[ii, "label2"] <- treeinfo2[jj, "node"]
-    ##                 pNode <- c(pNode, jp)
-    ##             }
-    ##             treeinfo[ii, "visited"] <- TRUE
-    ##         }
-    ##     }
-    ##     currentNode <- unique(pNode)
-    ## }
-    ## treeinfo[root, "label2"] <- root
-    ## ## convert nn to node that encoded in phylo2
-    ## node <- treeinfo$label2[match(nn, treeinfo$label)]
-
-
-    ####################################################
-    ##                                                ##
-    ##  after doing it in the hard way                ##
-    ##  I finally figure out the following easy way   ##
-    ##                                                ##
-    ####################################################
-
-    ## treeinfo <- fortify.phylo(phylo)
     tree_label <- c(phylo$tip.label, phylo$node.label)
     ii <- match(nn, tree_label)
 
@@ -196,14 +151,18 @@ read.stats_beast_internal <- function(beast, tree) {
     if (grepl("\\]:[0-9\\.eE+\\-]*\\[", tree) || grepl("\\]\\[", tree)) {
         ## MrBayes output
         stats <- strsplit(tree, "\\]:[0-9\\.eE+\\-]*\\[") %>% unlist
-        lstats <- lapply(stats, function(x) unlist(strsplit(x, split="\\][,\\)]")))
+        lstats <- lapply(stats, function(x) {
+            unlist(strsplit(x, split="\\][,\\)]"))
+        })
 
         for (i in seq_along(stats)) {
             n <- length(lstats[[i]])
             if (i == length(stats)) {
                 stats[i] <- lstats[[i]][n]
             } else {
-                stats[i] <- paste0(lstats[[i]][n], sub("&", ",", lstats[[i+1]][1]))
+                stats[i] <- paste0(lstats[[i]][n],
+                                   sub("&", ",", lstats[[i+1]][1])
+                                   )
             }
         }
         stats <- gsub("\\]\\[&", ",", stats)
@@ -230,13 +189,16 @@ read.stats_beast_internal <- function(beast, tree) {
         flag <- FALSE
         if (length(sidx) > 0) {
             flag <- TRUE
-            SETS <- sapply(seq_along(sidx), function(k) {
+            SETS <- lapply(seq_along(sidx), function(k) {
                 p <- y[sidx[k]:eidx[k]]
-                gsub(".*=\\{", "", p) %>% gsub("\\}$", "", .) %>% list
+                gsub(".*=\\{", "", p) %>% gsub("\\}$", "", .)
             })
             names(SETS) <- gsub("=.*", "", y[sidx])
 
-            kk <- sapply(seq_along(sidx), function(k) sidx[k]:eidx[k]) %>% unlist
+            kk <- lapply(seq_along(sidx), function(k) {
+                sidx[k]:eidx[k]
+            }) %>%
+                unlist
             y <- y[-kk]
         }
 
@@ -244,7 +206,8 @@ read.stats_beast_internal <- function(beast, tree) {
             return(SETS)
 
         name <- gsub("=.*", "", y)
-        val <- gsub(".*=", "", y) %>% gsub("^\\{", "", .) %>%
+        val <- gsub(".*=", "", y) %>%
+            gsub("^\\{", "", .) %>%
             gsub("\\}$", "", .)
 
         if (flag) {
@@ -262,7 +225,11 @@ read.stats_beast_internal <- function(beast, tree) {
         if (flag) {
             j <- i
             for (i in seq_along(SETS)) {
-                res[i+j] <- if(is_numeric(SETS[[i]])) list(as.numeric(SETS[[i]])) else SETS[i]
+                if(is_numeric(SETS[[i]])) {
+                    res[i+j] <- list(as.numeric(SETS[[i]]))
+                } else {
+                    res[i+j] <- SETS[i]
+                }
             }
         }
 
@@ -276,7 +243,7 @@ read.stats_beast_internal <- function(beast, tree) {
     stats2 <- lapply(stats2, function(x) {
         y <- x[nn]
         names(y) <- nn
-        y[sapply(y, is.null)] <- NA
+        y[vapply(y, is.null, logical(1))] <- NA
         y
     })
 
@@ -304,7 +271,10 @@ read.stats_beast_internal <- function(beast, tree) {
     colnames(stats3) <- cn
     stats3$node <- names(stats)
 
-    i <- sapply(stats3, function(x) max(sapply(x, length)))
+    i <- vapply(stats3,
+                function(x) max(vapply(x, length, numeric(1))),
+                numeric(1))
+
     for (j in which(i==1)) {
         stats3[,j] <- unlist(stats3[,j])
     }
@@ -318,7 +288,9 @@ add_pseudo_nodelabel <- function(phylo, treetext) {
         nnode <- phylo$Nnode
         nlab <- paste("X", 1:nnode, sep="")
         for (i in 1:nnode) {
-            treetext <- sub("\\)([:;])", paste0("\\)", nlab[i], "\\1"), treetext)
+            treetext <- sub("\\)([:;])",
+                            paste0("\\)", nlab[i], "\\1"),
+                            treetext)
         }
     }