+++ NEWS.md
@@ -1,3 +1,12 @@
+# treeio 1.3.11
+
++ fixed goodpractice check <2018-01-10, Wed>
+    - <https://github.com/ropensci/onboarding/issues/179#event-1416196637>
+    - avoid using = for assignment
+    - avoid code line > 80 characters
+    - avoid sapply, instead using vapply and lapply
+    - avoid using 1:length, 1:nrow and 1:ncol, use `seq_len` and `seq_along`
+
 # treeio 1.3.10
 
 * added 'Parsing jtree format' session in Importer vignette <2017-12-20, Wed>